fixing linter errors in form ui component
